Adding object for multiple cells
