+++ circusort/io/parameter/__init__.py
@@ -1,3 +1,4 @@
 from .base import generate_parameters, save_parameters, load_parameters, get_parameters
 from .data import load_data_parameters, get_data_parameters
-from .cell import load_cell_parameters, get_cell_parameters, get_cells_parameters
+from .cell import load_cell_parameters, get_cell_parameters
+from .cells import load_cells_parameters, get_cells_parameters
