+++ circusort/io/parameter/cell.py
@@ -1,7 +1,7 @@
 from .base import load_parameters, get_parameters
 
 
-default_cell_types = {
+default_types = {
     'template': {
         'mode': 'string',
         'path': 'string',
@@ -19,12 +19,6 @@ default_cell_types = {
     },
 }
 
-default_cells_types = {
-    'general': {
-        'nb_cells': 'integer',
-    },
-}
-
 
 def load_cell_parameters(path):
     """Load cell parameters from file.
@@ -38,7 +32,7 @@ def load_cell_parameters(path):
             The parameters of the cell.
     """
 
-    parameters = load_parameters(path, types=default_cell_types)
+    parameters = load_parameters(path, types=default_types)
 
     return parameters
 
@@ -55,23 +49,6 @@ def get_cell_parameters(path=None):
             The parameters of the cell.
     """
 
-    parameters = get_parameters(path=path, types=default_cell_types)
-
-    return parameters
-
-
-def get_cells_parameters(path=None):
-    """Get cells parameters from path.
-
-    Parameter:
-        path: none | string (optional)
-            The path to use to get the parameters of the cells. The default value is None.
-
-    Return:
-        parameters: circusort.obj.Parameters
-            The parameters of the cells.
-    """
-
-    parameters = get_parameters(path=path, types=default_cells_types, default_type='float')
+    parameters = get_parameters(path=path, types=default_types)
 
     return parameters
