+++ circusort/net/pregenerator.py
@@ -207,6 +207,9 @@ def pregenerator(working_directory=None, probe_path=None, parameters_path=None):
 
     cells = io.get_cells(probe=probe, **parameters['cells'])
     io.save_cells(generation_directory, cells)
+    # TODO enable the two following lines.
+    # cells.save(generation_directory)
+    # cells.plot(generation_directory)
     # TODO update the parameters of the cells.
     # parameters['cells'] = cells.get_parameters()
